requirements.txt
azure-core==1.38.0
azure-identity==1.25.1
azure-storage-blob==12.28.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
isodate==0.7.2
msal==1.34.0
msal-extensions==1.3.1
pycparser==3.0
PyJWT==2.10.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.3
